+++ CHANGELOG.md
@@ -12,6 +12,7 @@
         * Integer fields to store an integer in the given range.
         * Time fields which can store timestamps.
     * Experimental: Import strategies are experimental and may be removed in later versions.
+    * Renamed `FilterFieldTopN` function to `FilterAttrTopN`.
     * Removed all deprecated code.
     * Removed `Field` type and renamed `Frame` to `Field`.
     * Removed `client.ImportValueField` function. `client.ImportField` function imports both set and integer fields, depending on the type of the field.